+++ HISTORY.md
@@ -1,3 +1,8 @@
+Dev version
+-----------
+* Optimize redis driver `compose_message`.
+* Add module skynet.harbor for monitor harbor connect/disconnect, see test/testharborlink.lua .
+
 v0.3.1 (2014-6-16)
 -----------
 * Bugfix: lua mongo driver . Hold reply string before decode bson data.